Add files via upload
the game file contains the code given by the professor for the assaingment, and colors.json is a file that this code generates automatically but there is no funtionality for it at this point, that file is suposed to work with the cards that were not implemented in the end. Bot_Hard is my project and it needs the probabilidad.json to work properly, it's the data it uses to calculate the best move. I have'nt made jet the code for people to play against the bot.